Rename user endpoints to login endpoints and update corresponding service functions
