+++ app/api/services/login.py
@@ -1,22 +1,26 @@
-from app.core.db import SessionDep
+from sqlmodel import Session
 from app.api.models.user import User
 from sqlmodel import select
 from app.core.security import verify_password
 from fastapi import HTTPException
+from app.api.utils import send_email
+from app.core.security import create_access_token
+from app.core.config import settings
+from datetime import timedelta
 
 
-def get_user_by_id(session: SessionDep, email: str):
+def get_user_by_id(session: Session, email: str):
     statement = select(User).where(User.email == email)
     db_user = session.exec(statement).first()
     return db_user
 
 
-def get_all_users(session: SessionDep):
+def get_all_users(session: Session):
     statement = select(User)
     return session.exec(statement).all()
 
 
-def update_password(session: SessionDep, id: str, new_password):
+def update_password(session: Session, id: str, new_password):
     statement = select(User).where(User.id == id)
     user = session.exec(statement).first()
 
@@ -29,7 +33,7 @@ def update_password(session: SessionDep, id: str, new_password):
         raise HTTPException(status_code=404, detail=f"User with id {id} not found!")
 
 
-def authenticate(session: SessionDep, email: str, password: str):
+def authenticate(session: Session, email: str, password: str):
     db_user = get_user_by_id(session, email)
 
     if not db_user:
@@ -38,3 +42,30 @@ def authenticate(session: SessionDep, email: str, password: str):
         return None
     return db_user
 
+
+def dispatch_confirmation_email(user: User):
+    token = create_access_token(
+        subject=user.id,
+        expires_delta=timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES),
+    )
+    try:
+        link = f"{settings.URL_BASE}/login/confirm-email/{token}"
+        send_email(
+            email_to=user.email,
+            subject=f"{settings.PROJECT_NAME} - Confirme seu email",
+            content=f"Confirme seu email clicando nesse link: {link}",
+        )
+    except Exception as e:
+        raise e
+
+
+def dispatch_reset_password_email(email: str, token: str):
+    try:
+        link = f"{settings.URL_BASE}/reset-password/{token}"
+        send_email(
+            email_to=email,
+            subject=f"{settings.PROJECT_NAME} - Mude sua senha",
+            content=f"Confirme sua mudança de senha clicando nesse link: {link}",
+        )
+    except Exception as e:
+        raise e
